backend/requirements.txt
fastapi==0.110.0
python-multipart==0.0.9
uvicorn==0.27.1
aiofiles==23.2.1
python-magic-bin==0.4.14
yara-python==4.3.1
aiosqlite==0.19.0
filetype==1.2.0